requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.3.5
pypdf2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
numpy>=1.19.0,<2.0.0
qdrant-client==1.7.0
sentence-transformers>=3.0.0
aiofiles==23.2.1
httpx==0.25.2
